requirements.txt
anyio==3.6.1
attrs==22.1.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==3.4.8
ecdsa==0.18.0
fastapi==0.79.0
fastapi-jwt-auth==0.5.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.3
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2

jose~=1.0.0
emails~=0.6
alembic~=1.8.1